requirements.txt
ultralytics
numpy
streamlit
Pillow
keras
tensorflow